Rewrote all GET-handlers and added in the ORM-parts for permissions.
