+++ src/main/java/gov/polisen/ainwebserver/createHandlers/CREATEUser.java
@@ -6,15 +6,12 @@ import io.undertow.server.HttpServerExchange;
 
 import org.apache.ibatis.session.SqlSession;
 
-import com.google.gson.Gson;
-
 public class CREATEUser implements HttpHandler {
 
 	public void handleRequest(HttpServerExchange exchange) throws Exception {
 
 		SqlSession session = Main.getSessionFactory().openSession();
 		try {
-			Gson gson = new Gson();
 			// TODO: We need to first make a POST request from the app in order to
 			// test this data.
 		} finally {
