+++ src/main/java/gov/polisen/orm/clients/PermissionsOnCaseSqlProvider.java
@@ -1,0 +1,321 @@
+package gov.polisen.orm.clients;
+
+import static org.apache.ibatis.jdbc.SqlBuilder.BEGIN;
+import static org.apache.ibatis.jdbc.SqlBuilder.DELETE_FROM;
+import static org.apache.ibatis.jdbc.SqlBuilder.FROM;
+import static org.apache.ibatis.jdbc.SqlBuilder.INSERT_INTO;
+import static org.apache.ibatis.jdbc.SqlBuilder.ORDER_BY;
+import static org.apache.ibatis.jdbc.SqlBuilder.SELECT;
+import static org.apache.ibatis.jdbc.SqlBuilder.SELECT_DISTINCT;
+import static org.apache.ibatis.jdbc.SqlBuilder.SET;
+import static org.apache.ibatis.jdbc.SqlBuilder.SQL;
+import static org.apache.ibatis.jdbc.SqlBuilder.UPDATE;
+import static org.apache.ibatis.jdbc.SqlBuilder.VALUES;
+import static org.apache.ibatis.jdbc.SqlBuilder.WHERE;
+
+import gov.polisen.orm.examples.PermissionsOnCaseExample;
+import gov.polisen.orm.examples.PermissionsOnCaseExample.Criteria;
+import gov.polisen.orm.examples.PermissionsOnCaseExample.Criterion;
+import gov.polisen.orm.models.PermissionsOnCase;
+import java.util.List;
+import java.util.Map;
+
+public class PermissionsOnCaseSqlProvider {
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public String countByExample(PermissionsOnCaseExample example) {
+        BEGIN();
+        SELECT("count(*)");
+        FROM("permissionsoncases");
+        applyWhere(example, false);
+        return SQL();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public String deleteByExample(PermissionsOnCaseExample example) {
+        BEGIN();
+        DELETE_FROM("permissionsoncases");
+        applyWhere(example, false);
+        return SQL();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public String insertSelective(PermissionsOnCase record) {
+        BEGIN();
+        INSERT_INTO("permissionsoncases");
+        
+        if (record.getDeviceid() != null) {
+            VALUES("deviceid", "#{deviceid,jdbcType=INTEGER}");
+        }
+        
+        if (record.getCaseid() != null) {
+            VALUES("caseid", "#{caseid,jdbcType=INTEGER}");
+        }
+        
+        if (record.getUserid() != null) {
+            VALUES("userid", "#{userid,jdbcType=INTEGER}");
+        }
+        
+        if (record.getRead() != null) {
+            VALUES("read", "#{read,jdbcType=BIT}");
+        }
+        
+        if (record.getAdddata() != null) {
+            VALUES("adddata", "#{adddata,jdbcType=BIT}");
+        }
+        
+        if (record.getModify() != null) {
+            VALUES("modify", "#{modify,jdbcType=BIT}");
+        }
+        
+        if (record.getChangepermissions() != null) {
+            VALUES("changepermissions", "#{changepermissions,jdbcType=BIT}");
+        }
+        
+        return SQL();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public String selectByExample(PermissionsOnCaseExample example) {
+        BEGIN();
+        if (example != null && example.isDistinct()) {
+            SELECT_DISTINCT("deviceid");
+        } else {
+            SELECT("deviceid");
+        }
+        SELECT("caseid");
+        SELECT("userid");
+        SELECT("read");
+        SELECT("adddata");
+        SELECT("modify");
+        SELECT("changepermissions");
+        FROM("permissionsoncases");
+        applyWhere(example, false);
+        
+        if (example != null && example.getOrderByClause() != null) {
+            ORDER_BY(example.getOrderByClause());
+        }
+        
+        return SQL();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public String updateByExampleSelective(Map<String, Object> parameter) {
+        PermissionsOnCase record = (PermissionsOnCase) parameter.get("record");
+        PermissionsOnCaseExample example = (PermissionsOnCaseExample) parameter.get("example");
+        
+        BEGIN();
+        UPDATE("permissionsoncases");
+        
+        if (record.getDeviceid() != null) {
+            SET("deviceid = #{record.deviceid,jdbcType=INTEGER}");
+        }
+        
+        if (record.getCaseid() != null) {
+            SET("caseid = #{record.caseid,jdbcType=INTEGER}");
+        }
+        
+        if (record.getUserid() != null) {
+            SET("userid = #{record.userid,jdbcType=INTEGER}");
+        }
+        
+        if (record.getRead() != null) {
+            SET("read = #{record.read,jdbcType=BIT}");
+        }
+        
+        if (record.getAdddata() != null) {
+            SET("adddata = #{record.adddata,jdbcType=BIT}");
+        }
+        
+        if (record.getModify() != null) {
+            SET("modify = #{record.modify,jdbcType=BIT}");
+        }
+        
+        if (record.getChangepermissions() != null) {
+            SET("changepermissions = #{record.changepermissions,jdbcType=BIT}");
+        }
+        
+        applyWhere(example, true);
+        return SQL();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public String updateByExample(Map<String, Object> parameter) {
+        BEGIN();
+        UPDATE("permissionsoncases");
+        
+        SET("deviceid = #{record.deviceid,jdbcType=INTEGER}");
+        SET("caseid = #{record.caseid,jdbcType=INTEGER}");
+        SET("userid = #{record.userid,jdbcType=INTEGER}");
+        SET("read = #{record.read,jdbcType=BIT}");
+        SET("adddata = #{record.adddata,jdbcType=BIT}");
+        SET("modify = #{record.modify,jdbcType=BIT}");
+        SET("changepermissions = #{record.changepermissions,jdbcType=BIT}");
+        
+        PermissionsOnCaseExample example = (PermissionsOnCaseExample) parameter.get("example");
+        applyWhere(example, true);
+        return SQL();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public String updateByPrimaryKeySelective(PermissionsOnCase record) {
+        BEGIN();
+        UPDATE("permissionsoncases");
+        
+        if (record.getRead() != null) {
+            SET("read = #{read,jdbcType=BIT}");
+        }
+        
+        if (record.getAdddata() != null) {
+            SET("adddata = #{adddata,jdbcType=BIT}");
+        }
+        
+        if (record.getModify() != null) {
+            SET("modify = #{modify,jdbcType=BIT}");
+        }
+        
+        if (record.getChangepermissions() != null) {
+            SET("changepermissions = #{changepermissions,jdbcType=BIT}");
+        }
+        
+        WHERE("deviceid = #{deviceid,jdbcType=INTEGER}");
+        WHERE("caseid = #{caseid,jdbcType=INTEGER}");
+        WHERE("userid = #{userid,jdbcType=INTEGER}");
+        
+        return SQL();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    protected void applyWhere(PermissionsOnCaseExample example, boolean includeExamplePhrase) {
+        if (example == null) {
+            return;
+        }
+        
+        String parmPhrase1;
+        String parmPhrase1_th;
+        String parmPhrase2;
+        String parmPhrase2_th;
+        String parmPhrase3;
+        String parmPhrase3_th;
+        if (includeExamplePhrase) {
+            parmPhrase1 = "%s #{example.oredCriteria[%d].allCriteria[%d].value}";
+            parmPhrase1_th = "%s #{example.oredCriteria[%d].allCriteria[%d].value,typeHandler=%s}";
+            parmPhrase2 = "%s #{example.oredCriteria[%d].allCriteria[%d].value} and #{example.oredCriteria[%d].criteria[%d].secondValue}";
+            parmPhrase2_th = "%s #{example.oredCriteria[%d].allCriteria[%d].value,typeHandler=%s} and #{example.oredCriteria[%d].criteria[%d].secondValue,typeHandler=%s}";
+            parmPhrase3 = "#{example.oredCriteria[%d].allCriteria[%d].value[%d]}";
+            parmPhrase3_th = "#{example.oredCriteria[%d].allCriteria[%d].value[%d],typeHandler=%s}";
+        } else {
+            parmPhrase1 = "%s #{oredCriteria[%d].allCriteria[%d].value}";
+            parmPhrase1_th = "%s #{oredCriteria[%d].allCriteria[%d].value,typeHandler=%s}";
+            parmPhrase2 = "%s #{oredCriteria[%d].allCriteria[%d].value} and #{oredCriteria[%d].criteria[%d].secondValue}";
+            parmPhrase2_th = "%s #{oredCriteria[%d].allCriteria[%d].value,typeHandler=%s} and #{oredCriteria[%d].criteria[%d].secondValue,typeHandler=%s}";
+            parmPhrase3 = "#{oredCriteria[%d].allCriteria[%d].value[%d]}";
+            parmPhrase3_th = "#{oredCriteria[%d].allCriteria[%d].value[%d],typeHandler=%s}";
+        }
+        
+        StringBuilder sb = new StringBuilder();
+        List<Criteria> oredCriteria = example.getOredCriteria();
+        boolean firstCriteria = true;
+        for (int i = 0; i < oredCriteria.size(); i++) {
+            Criteria criteria = oredCriteria.get(i);
+            if (criteria.isValid()) {
+                if (firstCriteria) {
+                    firstCriteria = false;
+                } else {
+                    sb.append(" or ");
+                }
+                
+                sb.append('(');
+                List<Criterion> criterions = criteria.getAllCriteria();
+                boolean firstCriterion = true;
+                for (int j = 0; j < criterions.size(); j++) {
+                    Criterion criterion = criterions.get(j);
+                    if (firstCriterion) {
+                        firstCriterion = false;
+                    } else {
+                        sb.append(" and ");
+                    }
+                    
+                    if (criterion.isNoValue()) {
+                        sb.append(criterion.getCondition());
+                    } else if (criterion.isSingleValue()) {
+                        if (criterion.getTypeHandler() == null) {
+                            sb.append(String.format(parmPhrase1, criterion.getCondition(), i, j));
+                        } else {
+                            sb.append(String.format(parmPhrase1_th, criterion.getCondition(), i, j,criterion.getTypeHandler()));
+                        }
+                    } else if (criterion.isBetweenValue()) {
+                        if (criterion.getTypeHandler() == null) {
+                            sb.append(String.format(parmPhrase2, criterion.getCondition(), i, j, i, j));
+                        } else {
+                            sb.append(String.format(parmPhrase2_th, criterion.getCondition(), i, j, criterion.getTypeHandler(), i, j, criterion.getTypeHandler()));
+                        }
+                    } else if (criterion.isListValue()) {
+                        sb.append(criterion.getCondition());
+                        sb.append(" (");
+                        List<?> listItems = (List<?>) criterion.getValue();
+                        boolean comma = false;
+                        for (int k = 0; k < listItems.size(); k++) {
+                            if (comma) {
+                                sb.append(", ");
+                            } else {
+                                comma = true;
+                            }
+                            if (criterion.getTypeHandler() == null) {
+                                sb.append(String.format(parmPhrase3, i, j, k));
+                            } else {
+                                sb.append(String.format(parmPhrase3_th, i, j, k, criterion.getTypeHandler()));
+                            }
+                        }
+                        sb.append(')');
+                    }
+                }
+                sb.append(')');
+            }
+        }
+        
+        if (sb.length() > 0) {
+            WHERE(sb.toString());
+        }
+    }
+}
