+++ src/main/java/gov/polisen/orm/maps/PermissionsOnCaseMapper.java
@@ -1,0 +1,166 @@
+package gov.polisen.orm.maps;
+
+import gov.polisen.orm.clients.PermissionsOnCaseSqlProvider;
+import gov.polisen.orm.examples.PermissionsOnCaseExample;
+import gov.polisen.orm.models.PermissionsOnCase;
+import gov.polisen.orm.models.PermissionsOnCaseKey;
+import java.util.List;
+import org.apache.ibatis.annotations.Arg;
+import org.apache.ibatis.annotations.ConstructorArgs;
+import org.apache.ibatis.annotations.Delete;
+import org.apache.ibatis.annotations.DeleteProvider;
+import org.apache.ibatis.annotations.Insert;
+import org.apache.ibatis.annotations.InsertProvider;
+import org.apache.ibatis.annotations.Param;
+import org.apache.ibatis.annotations.Select;
+import org.apache.ibatis.annotations.SelectProvider;
+import org.apache.ibatis.annotations.Update;
+import org.apache.ibatis.annotations.UpdateProvider;
+import org.apache.ibatis.type.JdbcType;
+
+public interface PermissionsOnCaseMapper {
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @SelectProvider(type=PermissionsOnCaseSqlProvider.class, method="countByExample")
+    int countByExample(PermissionsOnCaseExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @DeleteProvider(type=PermissionsOnCaseSqlProvider.class, method="deleteByExample")
+    int deleteByExample(PermissionsOnCaseExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @Delete({
+        "delete from permissionsoncases",
+        "where deviceid = #{deviceid,jdbcType=INTEGER}",
+          "and caseid = #{caseid,jdbcType=INTEGER}",
+          "and userid = #{userid,jdbcType=INTEGER}"
+    })
+    int deleteByPrimaryKey(PermissionsOnCaseKey key);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @Insert({
+        "insert into permissionsoncases (deviceid, caseid, ",
+        "userid, read, adddata, ",
+        "modify, changepermissions)",
+        "values (#{deviceid,jdbcType=INTEGER}, #{caseid,jdbcType=INTEGER}, ",
+        "#{userid,jdbcType=INTEGER}, #{read,jdbcType=BIT}, #{adddata,jdbcType=BIT}, ",
+        "#{modify,jdbcType=BIT}, #{changepermissions,jdbcType=BIT})"
+    })
+    int insert(PermissionsOnCase record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @InsertProvider(type=PermissionsOnCaseSqlProvider.class, method="insertSelective")
+    int insertSelective(PermissionsOnCase record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @SelectProvider(type=PermissionsOnCaseSqlProvider.class, method="selectByExample")
+    @ConstructorArgs({
+        @Arg(column="deviceid", javaType=Integer.class, jdbcType=JdbcType.INTEGER, id=true),
+        @Arg(column="caseid", javaType=Integer.class, jdbcType=JdbcType.INTEGER, id=true),
+        @Arg(column="userid", javaType=Integer.class, jdbcType=JdbcType.INTEGER, id=true),
+        @Arg(column="read", javaType=Boolean.class, jdbcType=JdbcType.BIT),
+        @Arg(column="adddata", javaType=Boolean.class, jdbcType=JdbcType.BIT),
+        @Arg(column="modify", javaType=Boolean.class, jdbcType=JdbcType.BIT),
+        @Arg(column="changepermissions", javaType=Boolean.class, jdbcType=JdbcType.BIT)
+    })
+    List<PermissionsOnCase> selectByExample(PermissionsOnCaseExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @Select({
+        "select",
+        "deviceid, caseid, userid, read, adddata, modify, changepermissions",
+        "from permissionsoncases",
+        "where deviceid = #{deviceid,jdbcType=INTEGER}",
+          "and caseid = #{caseid,jdbcType=INTEGER}",
+          "and userid = #{userid,jdbcType=INTEGER}"
+    })
+    @ConstructorArgs({
+        @Arg(column="deviceid", javaType=Integer.class, jdbcType=JdbcType.INTEGER, id=true),
+        @Arg(column="caseid", javaType=Integer.class, jdbcType=JdbcType.INTEGER, id=true),
+        @Arg(column="userid", javaType=Integer.class, jdbcType=JdbcType.INTEGER, id=true),
+        @Arg(column="read", javaType=Boolean.class, jdbcType=JdbcType.BIT),
+        @Arg(column="adddata", javaType=Boolean.class, jdbcType=JdbcType.BIT),
+        @Arg(column="modify", javaType=Boolean.class, jdbcType=JdbcType.BIT),
+        @Arg(column="changepermissions", javaType=Boolean.class, jdbcType=JdbcType.BIT)
+    })
+    PermissionsOnCase selectByPrimaryKey(PermissionsOnCaseKey key);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @UpdateProvider(type=PermissionsOnCaseSqlProvider.class, method="updateByExampleSelective")
+    int updateByExampleSelective(@Param("record") PermissionsOnCase record, @Param("example") PermissionsOnCaseExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @UpdateProvider(type=PermissionsOnCaseSqlProvider.class, method="updateByExample")
+    int updateByExample(@Param("record") PermissionsOnCase record, @Param("example") PermissionsOnCaseExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @UpdateProvider(type=PermissionsOnCaseSqlProvider.class, method="updateByPrimaryKeySelective")
+    int updateByPrimaryKeySelective(PermissionsOnCase record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    @Update({
+        "update permissionsoncases",
+        "set read = #{read,jdbcType=BIT},",
+          "adddata = #{adddata,jdbcType=BIT},",
+          "modify = #{modify,jdbcType=BIT},",
+          "changepermissions = #{changepermissions,jdbcType=BIT}",
+        "where deviceid = #{deviceid,jdbcType=INTEGER}",
+          "and caseid = #{caseid,jdbcType=INTEGER}",
+          "and userid = #{userid,jdbcType=INTEGER}"
+    })
+    int updateByPrimaryKey(PermissionsOnCase record);
+}
