--- src/main/java/gov/polisen/orm/maps/PermissionsOnCasesMapper.java
@@ -1,132 +1,0 @@
-package gov.polisen.orm.maps;
-
-import gov.polisen.orm.clients.PermissionsOnCasesSqlProvider;
-import gov.polisen.orm.examples.PermissionsOnCasesExample;
-import gov.polisen.orm.models.PermissionsOnCases;
-import gov.polisen.orm.models.PermissionsOnCasesKey;
-import java.util.List;
-import org.apache.ibatis.annotations.Arg;
-import org.apache.ibatis.annotations.ConstructorArgs;
-import org.apache.ibatis.annotations.Delete;
-import org.apache.ibatis.annotations.DeleteProvider;
-import org.apache.ibatis.annotations.Insert;
-import org.apache.ibatis.annotations.InsertProvider;
-import org.apache.ibatis.annotations.Param;
-import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.SelectProvider;
-import org.apache.ibatis.annotations.Update;
-import org.apache.ibatis.annotations.UpdateProvider;
-import org.apache.ibatis.type.JdbcType;
-
-public interface PermissionsOnCasesMapper {
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@SelectProvider(type = PermissionsOnCasesSqlProvider.class, method = "countByExample")
-	int countByExample(PermissionsOnCasesExample example);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@DeleteProvider(type = PermissionsOnCasesSqlProvider.class, method = "deleteByExample")
-	int deleteByExample(PermissionsOnCasesExample example);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@Delete({ "delete from permissionsondeployments",
-			"where deploymentid = #{deploymentid,jdbcType=INTEGER}",
-			"and userid = #{userid,jdbcType=INTEGER}" })
-	int deleteByPrimaryKey(PermissionsOnCasesKey key);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@Insert({
-			"insert into permissionsondeployments (deploymentid, userid, ",
-			"read, adddata, modify, ",
-			"changepermissions)",
-			"values (#{deploymentid,jdbcType=INTEGER}, #{userid,jdbcType=INTEGER}, ",
-			"#{read,jdbcType=BIT}, #{adddata,jdbcType=BIT}, #{modify,jdbcType=BIT}, ",
-			"#{changepermissions,jdbcType=BIT})" })
-	int insert(PermissionsOnCases record);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@InsertProvider(type = PermissionsOnCasesSqlProvider.class, method = "insertSelective")
-	int insertSelective(PermissionsOnCases record);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@SelectProvider(type = PermissionsOnCasesSqlProvider.class, method = "selectByExample")
-	@ConstructorArgs({
-			@Arg(column = "deploymentid", javaType = Integer.class, jdbcType = JdbcType.INTEGER, id = true),
-			@Arg(column = "userid", javaType = Integer.class, jdbcType = JdbcType.INTEGER, id = true),
-			@Arg(column = "read", javaType = Boolean.class, jdbcType = JdbcType.BIT),
-			@Arg(column = "adddata", javaType = Boolean.class, jdbcType = JdbcType.BIT),
-			@Arg(column = "modify", javaType = Boolean.class, jdbcType = JdbcType.BIT),
-			@Arg(column = "changepermissions", javaType = Boolean.class, jdbcType = JdbcType.BIT) })
-	List<PermissionsOnCases> selectByExample(PermissionsOnCasesExample example);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@Select({ "select",
-			"deploymentid, userid, read, adddata, modify, changepermissions",
-			"from permissionsondeployments",
-			"where deploymentid = #{deploymentid,jdbcType=INTEGER}",
-			"and userid = #{userid,jdbcType=INTEGER}" })
-	@ConstructorArgs({
-			@Arg(column = "deploymentid", javaType = Integer.class, jdbcType = JdbcType.INTEGER, id = true),
-			@Arg(column = "userid", javaType = Integer.class, jdbcType = JdbcType.INTEGER, id = true),
-			@Arg(column = "read", javaType = Boolean.class, jdbcType = JdbcType.BIT),
-			@Arg(column = "adddata", javaType = Boolean.class, jdbcType = JdbcType.BIT),
-			@Arg(column = "modify", javaType = Boolean.class, jdbcType = JdbcType.BIT),
-			@Arg(column = "changepermissions", javaType = Boolean.class, jdbcType = JdbcType.BIT) })
-	PermissionsOnCases selectByPrimaryKey(PermissionsOnCasesKey key);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@UpdateProvider(type = PermissionsOnCasesSqlProvider.class, method = "updateByExampleSelective")
-	int updateByExampleSelective(@Param("record") PermissionsOnCases record,
-			@Param("example") PermissionsOnCasesExample example);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@UpdateProvider(type = PermissionsOnCasesSqlProvider.class, method = "updateByExample")
-	int updateByExample(@Param("record") PermissionsOnCases record,
-			@Param("example") PermissionsOnCasesExample example);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@UpdateProvider(type = PermissionsOnCasesSqlProvider.class, method = "updateByPrimaryKeySelective")
-	int updateByPrimaryKeySelective(PermissionsOnCases record);
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	@Update({ "update permissionsondeployments",
-			"set read = #{read,jdbcType=BIT},", "adddata = #{adddata,jdbcType=BIT},",
-			"modify = #{modify,jdbcType=BIT},",
-			"changepermissions = #{changepermissions,jdbcType=BIT}",
-			"where deploymentid = #{deploymentid,jdbcType=INTEGER}",
-			"and userid = #{userid,jdbcType=INTEGER}" })
-	int updateByPrimaryKey(PermissionsOnCases record);
-}
