+++ src/main/java/gov/polisen/orm/models/PermissionsOnCaseKey.java
@@ -1,0 +1,75 @@
+package gov.polisen.orm.models;
+
+public class PermissionsOnCaseKey {
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column permissionsoncases.deviceid
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    private Integer deviceid;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column permissionsoncases.caseid
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    private Integer caseid;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column permissionsoncases.userid
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    private Integer userid;
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table permissionsoncases
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public PermissionsOnCaseKey(Integer deviceid, Integer caseid, Integer userid) {
+        this.deviceid = deviceid;
+        this.caseid = caseid;
+        this.userid = userid;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column permissionsoncases.deviceid
+     *
+     * @return the value of permissionsoncases.deviceid
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public Integer getDeviceid() {
+        return deviceid;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column permissionsoncases.caseid
+     *
+     * @return the value of permissionsoncases.caseid
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public Integer getCaseid() {
+        return caseid;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column permissionsoncases.userid
+     *
+     * @return the value of permissionsoncases.userid
+     *
+     * @mbggenerated Sat Apr 05 19:09:46 CEST 2014
+     */
+    public Integer getUserid() {
+        return userid;
+    }
+}
