--- src/main/java/gov/polisen/orm/models/PermissionsOnCases.java
@@ -1,74 +1,0 @@
-package gov.polisen.orm.models;
-
-public class PermissionsOnCases extends PermissionsOnCasesKey {
-
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column permissionsondeployments.read
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	private Boolean	read;
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column permissionsondeployments.adddata
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	private Boolean	adddata;
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column permissionsondeployments.modify
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	private Boolean	modify;
-	/**
-	 * This field was generated by MyBatis Generator. This field corresponds to the database column permissionsondeployments.changepermissions
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	private Boolean	changepermissions;
-
-	/**
-	 * This method was generated by MyBatis Generator. This method corresponds to the database table permissionsondeployments
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	public PermissionsOnCases(Integer deploymentid, Integer userid, Boolean read,
-			Boolean adddata, Boolean modify, Boolean changepermissions) {
-		super(deploymentid, userid);
-		this.read = read;
-		this.adddata = adddata;
-		this.modify = modify;
-		this.changepermissions = changepermissions;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column permissionsondeployments.read
-	 * @return  the value of permissionsondeployments.read
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	public Boolean getRead() {
-		return read;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column permissionsondeployments.adddata
-	 * @return  the value of permissionsondeployments.adddata
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	public Boolean getAdddata() {
-		return adddata;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column permissionsondeployments.modify
-	 * @return  the value of permissionsondeployments.modify
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	public Boolean getModify() {
-		return modify;
-	}
-
-	/**
-	 * This method was generated by MyBatis Generator. This method returns the value of the database column permissionsondeployments.changepermissions
-	 * @return  the value of permissionsondeployments.changepermissions
-	 * @mbggenerated  Sat Apr 05 11:34:16 CEST 2014
-	 */
-	public Boolean getChangepermissions() {
-		return changepermissions;
-	}
-}
